Update patch set 1
Patch Set 1:

Does fdlibm need to be built with multimode support? Ie the ability to change the behavior of the library by assigning _LIB_VERSION?

The comment in the makefile suggests that the library only needs to support IEE754 mode and it tries to achieve this by defining _LIB_VERSION as "const enum fdversion _IEEE_" on the command line, but this does nothing as fdlibm.h unconditionally redefines _LIB_VERSION to be "enum fdversion".

Without _IEEE_LIBM being defined, every routine contains a runtime check of _LIB_VERSION. If the command line definition of _LIB_VERSION had worked as intended the compiler would  treat _LIB_VERSION as constant and optimise the code to make it equivalent to defining _IEEE_LIBM
Eg from w_sinh.c (slightly simplified):

 double ieee_sinh(double x)              /* wrapper sinh */
 {
 #ifdef _IEEE_LIBM
        return __ieee754_sinh(x);
 #else
        double z; 
        z = __ieee754_sinh(x);
        if(_LIB_VERSION == _IEEE_) return z;
        if(!ieee_finite(z)&&ieee_finite(x)) {
            return __kernel_standard(x,x,25); /* sinh overflow */
        } else
            return z;
 #endif
 }

Something needs to change in Android.mk. Either it only needs to support IEEE754 mode:

  # This guarantees that the FDLIBM functions are only in
  # "IEEE spirit", i.e. only the IEEE 754 core functions
  # are used.
 cflags := -D_IEEE_LIBM

or it needs to support multimode:

  # This sets the default FDLIBM mode to IEEE754. At
  # runtime the library may be reconfigured to POSIX, XOPEN
  # or SVID3 mode by assigning _LIB_VERSION
  cflags := -D_IEEE_MODE

If you can give some guidance on the expected behavior of fdlibm I can update the patch to one of the above choices.

I'm not sure how to make a test case for this. You mentioned previously that there had been some problems with StrictMath tests? Could you send me the context and maybe I could come up with something?

Thanks

Patch-set: 1