Update patch set 1
Patch Set 1:

I found the commit that changed from the cflags command line. It was to get correct operation of dalvikvm on x86 host.
da0692bd3a2a146aa7ea5ce30002c584761d5ad0

I wonder if forcing the library to do runtime checks worked around some of the problems that the later no-strict-aliasing change fixed more comprehensively?
980942a285dffed1cfe323d46444e40074868dff

For MIPS we use no-strict-aliasing globally. I know that one of the reasons we did this was to fix problems in fdlibm and at the time we weren't sure if there would be problems elsewhere in the tree.

Patch-set: 1